update neon
AVXはテストNG
