+++ dSFMT-avx2.h
@@ -1,11 +1,26 @@
 #pragma once
-
-#ifndef DSFMT_X86AVX_H
-#define DSFMT_X86AVX_H
+/**
+ * @file dSFMT-avx2.h
+ *
+ * @brief SIMD oriented Fast Mersenne Twister(SFMT) pseudorandom
+ * number generator with jump function. This file includes common functions
+ * used in random number generation and jump.
+ *
+ * @author Masaki Ota
+ *
+ * Copyright (C) 2025 Masaki Ota.
+ *
+ * The 3-clause BSD License is applied to this software, see
+ * LICENSE.txt
+ */
+#ifndef DSFMT_AVX2_H
+#define DSFMT_AVX2_H
 
 #include <immintrin.h>
 #include <emmintrin.h>
 #include "dSFMT-params.h"
+
+#define DSFMT_RECURSION_X2
 
 /**
  * This function represents the recursion formula.
@@ -17,7 +32,7 @@
 inline static __m256i do_recursion_x2_avx2(__m256i a, __m256i b, __m256i *u) {
     __m256i v, w, x, y, z;
     __m256i lung_shuf_idx = _mm256_setr_epi32(7, 6, 5, 4, 4, 5, 6, 7);
-    __m256i word_reverse_idx = _mm256_setr_epi32(0, 1, 2, 3, 4, 5, 6, 7);
+    __m256i word_reverse_idx = _mm256_setr_epi32(0, 1, 2, 3, 3, 2, 1, 0);
     __m256i z_mask = _mm256_setr_epi32(0, 0, 0, 0, -1, -1, -1, -1);
     __m256i dsfmt_mask = _mm256_broadcastsi128_si256(_mm_set_epi64x(DSFMT_MSK2, DSFMT_MSK1));
 
@@ -40,7 +55,7 @@ inline static __m256i do_recursion_x2_avx2(__m256i a, __m256i b, __m256i *u) {
     w = _mm256_and_si256(y, dsfmt_mask);
     v = _mm256_xor_si256(v, x);
     v = _mm256_xor_si256(v, w);
-    _mm256_store_si256((__m256i*)r, v);
+    return v;
 }
 
 /**
@@ -68,34 +83,27 @@ inline static __m128i do_recursion_last(__m128i a, __m128i b, __m128i* u) {
     return v;
 }
 
-
-
-//#define DO_RECURSION_X4
-//#define DO_RECURSION_X3
-#define DO_RECURSION_X2
-
-
 typedef void (*avx_converter_t)(__m256d *);
 
-inline static void* no_convert_avx(__m256d* w) {}
-inline static void* convert_c0o1_avx(__m256d* w) {
-    *w = _mm256_add_pd(w, _mm256_set1_pd(-1.0));
+inline static void no_convert_avx(__m256d* w) {}
+inline static void convert_c0o1_avx(__m256d* w) {
+    *w = _mm256_add_pd(*w, _mm256_set1_pd(-1.0));
 }
 
-inline static void* convert_o0c1_avx(__m256d* w) {
-    *w = _mm256_sub_pd(_mm256_set1_pd(2.0), w);
+inline static void convert_o0c1_avx(__m256d* w) {
+    *w = _mm256_sub_pd(_mm256_set1_pd(2.0), *w);
 }
-inline static void* convert_o0o1_avx(__m256d* w) {
+inline static void convert_o0o1_avx(__m256d* w) {
     __m256d mask = _mm256_castsi256_pd(_mm256_set1_epi64x(1));
-    *w = _mm256_add_pd(_mm256_or_pd(w, mask), _mm256_set1_pd(-1.0));
+    *w = _mm256_add_pd(_mm256_or_pd(*w, mask), _mm256_set1_pd(-1.0));
 }
 
-#define MAKE_PAIR(a, b) _mm256_set_m128i((b).si, (a).si)
+#define MAKE_PAIR(a, b) _mm256_set_m128i((b)->si, (a)->si)
 #define LD256(a) _mm256_loadu_si256((__m256i*)(a))
 #define LD128(a) _mm_loadu_si128((__m128i*)(a))
 
-#define ST256(addr, v) _mm256_storeu_si256(((__m256i*)(addr), (v));
-#define ST128(addr, v) _mm_storeu_si128(((__m256i*)(addr), (v));
+#define ST256(addr, v) _mm256_storeu_si256((__m256i*)(addr), (v));
+#define ST128(addr, v) _mm_storeu_si128((__m128i*)(addr), (v));
 
 /**
  * This is the core function for generating pseudorandom numbers.
@@ -113,7 +121,7 @@ inline static void gen_rand_array_avx_core(dsfmt_t *dsfmt, w128_t *array,
     __m256i lung = _mm256_broadcastsi128_si256(dsfmt->status[DSFMT_N].si);
 
     const ptrdiff_t loop1_end = DSFMT_N - DSFMT_POS1;
-    for (; i < loop1_end - 1; i += 2) {
+    for (i=0; i < loop1_end - 1; i += 2) {
         __m256i result = do_recursion_x2_avx2(
             LD256(&dsfmt->status[i]),
             LD256(&dsfmt->status[i + DSFMT_POS1]), 
@@ -154,13 +162,13 @@ inline static void gen_rand_array_avx_core(dsfmt_t *dsfmt, w128_t *array,
             LD256(&array[i + DSFMT_POS1 - DSFMT_N]),
             &lung);
         ST256(&array[i], result);
-        converter(&array[i - DSFMT_N]);
+        converter((__m256d*)&array[i - DSFMT_N]);
     }
 
     // ループ4: 次の状態ベクトルを準備 (前半)
     const ptrdiff_t loop4_end = 2 * DSFMT_N - size;
     for (j = 0; j < loop4_end-1; j+=2) {
-        ST256(&dsfmt->status[j], LD256[&array[j + size - DSFMT_N]]);
+        ST256(&dsfmt->status[j], LD256(&array[j + size - DSFMT_N]));
     }
 
     // ループ5: 配列の残りを生成しつつ、次の状態ベクトルを準備 (後半)
@@ -170,7 +178,7 @@ inline static void gen_rand_array_avx_core(dsfmt_t *dsfmt, w128_t *array,
             LD256(&array[i + DSFMT_POS1 - DSFMT_N]),
             &lung);
         ST256(&array[i], result);
-        converter(&array[i - DSFMT_N]);
+        converter((__m256d*)&array[i - DSFMT_N]);
     }
     if (size & 1) {
         __m128i lung128 = _mm256_extracti128_si256(lung, 1);
@@ -186,21 +194,22 @@ inline static void gen_rand_array_avx_core(dsfmt_t *dsfmt, w128_t *array,
 
     // ループ6: 生成された配列の末尾部分に変換を適用
     for (i = loop3_end; i < size; i+= 2) {
-        converter(&array[i]);
+        converter((__m256d*)&array[i]);
     }
     if (size & 1) {
-        __m256d data = _mm256_castpd128_pd256(_mm_loadu_ps(&array[i]))
+        __m256d data = _mm256_castpd128_pd256(_mm_loadu_pd((double*)&array[i]));
         converter(&data);
         _mm_storeu_pd((double*)&array[i], _mm256_castpd256_pd128(data));
     } 
 }
+
 
 /**
  * This function fills the internal state array with double precision
  * floating point pseudorandom numbers of the IEEE 754 format.
  * @param dsfmt dsfmt state vector.
  */
-void dsfmt_gen_rand_all_avx(dsfmt_t *dsfmt) {
+void dsfmt_gen_rand_all(dsfmt_t *dsfmt) {
     int i;
     __m256i lung = _mm256_broadcastsi128_si256(dsfmt->status[DSFMT_N].si);
 
@@ -232,20 +241,18 @@ void dsfmt_gen_rand_all_avx(dsfmt_t *dsfmt) {
     }
     if (DSFMT_N & 1) {
         __m128i lung128 = _mm256_extracti128_si256(lung, 1);
-        __m128i result = do_recursion_avx2(
-            LD256(&dsfmt->status[DSFMT_N-1]), 
-            LD256(&array[i + DSFMT_POS1 - DSFMT_N]),
+        __m128i result = do_recursion_last(
+            LD128(&dsfmt->status[DSFMT_N-1]), 
+            LD128(&dsfmt->status[i + DSFMT_POS1 - DSFMT_N]),
             &lung128);
         ST128(&dsfmt->status[DSFMT_N-1], result);
         ST128(&dsfmt->status[DSFMT_N], lung128);
-
-    } else {
-                ST128(&dsfmt->status[DSFMT_N], lung128);
     } else {
         ST128(&dsfmt->status[DSFMT_N], _mm256_extracti128_si256(lung, 1));
     }
 
 }
+#define DSFMT_GEN_DRAND_ALL
 
-#endif /* DSFMT_ARM64NEON_H */
+#endif /* DSFMT_X86AVX_H */
