+++ dSFMT-common.h
@@ -39,13 +39,11 @@ union X128D_T {
 /** mask data for sse2 */
 static const union X128I_T sse2_param_mask = {{DSFMT_MSK1, DSFMT_MSK2}};
 
-#elif defined(__aarch64__)
-  #if !defined(ARM64_NO_NEON)
-    #define HAVE_NEON
-    #if defined(__ARM_FEATURE_SHA3) && !defined(ARM64_NO_SHA3)
-      #define HAVE_SHA3
-    #endif 
-  #endif 
+#elif defined(__aarch64__) && !defined(ARM64_NO_NEON)
+#define HAVE_NEON
+#if defined(__ARM_FEATURE_SHA3) 
+#define HAVE_SHA3
+#endif 
 #endif
 
 #if defined(HAVE_ALTIVEC)
@@ -77,10 +75,10 @@ inline static void do_recursion(w128_t *r, w128_t *a, w128_t * b,
     r->s = vec_xor(z, x);
     lung->s = w;
 }
+#elif defined(HAVE_AVX2) && DSFMT_MEXP >= 2203
+#include "dSFMT-avx2.h"
+#define DSFMT_AVX2_DECLEARED
 #elif defined(HAVE_SSE2)
-#if defined(HAVE_AVX2) && (DSFMT_N - DSFMT_POS >= 2)
-#include "dSFMT-x86avx.h"
-#else
 /**
  * This function represents the recursion formula.
  * @param r output 128-bit
@@ -112,9 +110,54 @@ inline static void do_recursion(w128_t *r, w128_t *a, w128_t *b, w128_t *u) {
     u->si = y;
 }
 
+inline static void do_recursion_x2(w128_t *r, w128_t *a, w128_t *b, w128_t *u) {
+    __m128i v1, w1, x1, y1, z1;
+    __m128i v2, w2, x2, y2, z2;
 
+    x1 = a[0].si;
+    x2 = a[1].si;
+
+    z1 = _mm_slli_epi64(x1, DSFMT_SL1);
+    z2 = _mm_slli_epi64(x2, DSFMT_SL1);
+
+    z1 = _mm_xor_si128(z1, b[0].si);
+    z2 = _mm_xor_si128(z2, b[1].si);
+    
+    v1 = _mm_shuffle_epi32(z1, SSE2_SHUFF);
+    z2 = _mm_xor_si128(z2, v1);
+
+    y2 = u->si;
+    y1 = _mm_shuffle_epi32(y2, SSE2_SHUFF);
+    y1 = _mm_xor_si128(y1, z1);
+    y2 = _mm_xor_si128(y2, z2);
+    u->si = y2;
+
+    v1 = _mm_srli_epi64(y1, DSFMT_SR);
+    v2 = _mm_srli_epi64(y2, DSFMT_SR);
+    w1 = _mm_and_si128(y1, sse2_param_mask.i128);
+    w2 = _mm_and_si128(y2, sse2_param_mask.i128);
+    v1 = _mm_xor_si128(v1, x1);
+    v2 = _mm_xor_si128(v2, x2);
+    v1 = _mm_xor_si128(v1, w1);
+    v2 = _mm_xor_si128(v2, w2);
+
+    r[0].si = v1;
+    r[1].si = v2;
+}
+
+inline static void do_recursion_x3(w128_t *r, w128_t *a, w128_t *b, w128_t *u) {
+    do_recursion_x2(r, a, b, u);
+    do_recursion(r+2, a+2, b+2, u);
+}
+
+#if DSFMT_MEXP == 1279
+#define DSFMT_RECURSION_X3
+#elif DSFMT_N - DSFMT_POS1 > 1 
+#define DSFMT_RECURSION_X2
+#endif
+ 
 #elif defined(HAVE_NEON)
-#include "dSFMT-arm64neon.h"
+#include "dSFMT-neon.h"
 #else
 /**
  * This function represents the recursion formula.
