+++ dSFMT.h
@@ -163,6 +163,7 @@ union W128_T {
     uint64x2_t u64x2;
     uint32x4_t u32x4;
     uint8x16_t u8x16;
+    
     float64x2_t f64x2;
 
     uint64_t u[2];
